Add CurrentConnection Property into Interfaceinfo. this property value will change each call
